LayoutTests/http/wpt/wk-web-animations/timing-model/timeline-current-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Check that the AnimationTimeline currentTime property is defined and can be set via internals</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

internals.pauseTimeline(document.timeline);

test(t => {
    assert_equals(document.timeline.currentTime, null);
}, 'The document timeline currentTime is null by default.');

test(t => {
    internals.setTimelineCurrentTime(document.timeline, 1);
    assert_equals(document.timeline.currentTime, 1);
}, 'Using internals.setTimelineCurrentTime() can set the document timeline currentTime.');

test(t => {
    internals.setTimelineCurrentTime(document.timeline, -1);
    assert_equals(document.timeline.currentTime, -1);
}, 'Using internals.setTimelineCurrentTime() can set the document timeline currentTime to a negative value.');

</script>
</body>
